dedup some test config
